// gql/graphql.ts
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type AboutModelDescriptionField = {
  __typename?: 'AboutModelDescriptionField';
  blocks: Array<Scalars['String']['output']>;
  links: Array<Scalars['String']['output']>;
  value: Scalars['JsonField']['output'];
};
// ...
export type AboutModelMalgosiaDescriptionField = {
  __typename?: 'AboutModelMalgosiaDescriptionField';
  blocks: Array<Scalars['String']['output']>;
  links: Array<Scalars['String']['output']>;
  value: Scalars['JsonField']['output'];
};
// ...
export type AboutRecord = RecordInterface & {
  __typename?: 'AboutRecord';
  _allDescriptionLocales?: Maybe<Array<AboutModelDescriptionFieldMultiLocaleField>>;
  _allTitleLocales?: Maybe<Array<StringMultiLocaleField>>;
  _createdAt: Scalars['DateTime']['output'];
  /** Editing URL */
  _editingUrl?: Maybe<Scalars['String']['output']>;
  _firstPublishedAt?: Maybe<Scalars['DateTime']['output']>;
  _isValid: Scalars['BooleanType']['output'];
  _locales: Array<SiteLocale>;
  _modelApiKey: Scalars['String']['output'];
  _publicationScheduledAt?: Maybe<Scalars['DateTime']['output']>;
  _publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** Generates SEO and Social card meta tags to be used in your frontend */
  _seoMetaTags: Array<Tag>;
  _status: ItemStatus;
  _unpublishingScheduledAt?: Maybe<Scalars['DateTime']['output']>;
  _updatedAt: Scalars['DateTime']['output'];
  description?: Maybe<AboutModelDescriptionField>;
  id: Scalars['ItemId']['output'];
  malgosiaDescription?: Maybe<AboutModelMalgosiaDescriptionField>;
  malgosiaImage?: Maybe<FileField>;
  title?: Maybe<Scalars['String']['output']>;
};
// ...
export type FileField = FileFieldInterface & {
  __typename?: 'FileField';
  _createdAt: Scalars['DateTime']['output'];
  /** The DatoCMS URL where you can edit this entity. To use this field, you need to set a X-Base-Editing-Url header in the request */
  _editingUrl?: Maybe<Scalars['String']['output']>;
  _updatedAt: Scalars['DateTime']['output'];
  alt?: Maybe<Scalars['String']['output']>;
  author?: Maybe<Scalars['String']['output']>;
  basename: Scalars['String']['output'];
  blurUpThumb?: Maybe<Scalars['String']['output']>;
  blurhash?: Maybe<Scalars['String']['output']>;
  colors: Array<ColorField>;
  copyright?: Maybe<Scalars['String']['output']>;
  customData: Scalars['CustomData']['output'];
  exifInfo: Scalars['CustomData']['output'];
  filename: Scalars['String']['output'];
  focalPoint?: Maybe<FocalPoint>;
  format: Scalars['String']['output'];
  height?: Maybe<Scalars['IntType']['output']>;
  id: Scalars['UploadId']['output'];
  md5: Scalars['String']['output'];
  mimeType: Scalars['String']['output'];
  notes?: Maybe<Scalars['String']['output']>;
  responsiveImage?: Maybe<ResponsiveImage>;
  size: Scalars['IntType']['output'];
  smartTags: Array<Scalars['String']['output']>;
  tags: Array<Scalars['String']['output']>;
  thumbhash?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
  url: Scalars['String']['output'];
  video?: Maybe<UploadVideoField>;
  width?: Maybe<Scalars['IntType']['output']>;
};
// ...
export type PersonRecord = RecordInterface & {
  __typename?: 'PersonRecord';
  _createdAt: Scalars['DateTime']['output'];
  /** Editing URL */
  _editingUrl?: Maybe<Scalars['String']['output']>;
  _firstPublishedAt?: Maybe<Scalars['DateTime']['output']>;
  _isValid: Scalars['BooleanType']['output'];
  _modelApiKey: Scalars['String']['output'];
  _publicationScheduledAt?: Maybe<Scalars['DateTime']['output']>;
  _publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** Generates SEO and Social card meta tags to be used in your frontend */
  _seoMetaTags: Array<Tag>;
  _status: ItemStatus;
  _unpublishingScheduledAt?: Maybe<Scalars['DateTime']['output']>;
  _updatedAt: Scalars['DateTime']['output'];
  id: Scalars['ItemId']['output'];
  image?: Maybe<FileField>;
  name?: Maybe<Scalars['String']['output']>;
  role?: Maybe<Scalars['String']['output']>;
};
// ...
export type Query = {
  __typename?: 'Query';
  /** Returns meta information regarding a record collection */
  _allAboutsMeta: CollectionMetadata;
  /** Returns meta information regarding a record collection */
  _allEventsMeta: CollectionMetadata;
  /** Returns meta information regarding a record collection */
  _allGalleriesMeta: CollectionMetadata;
  /** Returns meta information regarding a record collection */
  _allJoinsMeta: CollectionMetadata;
  /** Returns meta information regarding a record collection */
  _allPartnersMeta: CollectionMetadata;
  /** Returns meta information regarding a record collection */
  _allPodcastsMeta: CollectionMetadata;
  /** Returns meta information regarding an assets collection */
  _allUploadsMeta: CollectionMetadata;
  /** Returns the single instance record */
  _site: Site;
  /** Returns a specific record */
  about?: Maybe<AboutRecord>;
  /** Returns a collection of records */
  allAbouts: Array<AboutRecord>;
  /** Returns a collection of records */
  allEvents: Array<EventRecord>;
  /** Returns a collection of records */
  allGalleries: Array<GalleryRecord>;
  /** Returns a collection of records */
  allJoins: Array<JoinRecord>;
  /** Returns a collection of records */
  allPartners: Array<PartnerRecord>;
  /** Returns a collection of records */
  allPodcasts: Array<PodcastRecord>;
  /** Returns a collection of assets */
  allUploads: Array<FileField>;
  /** Returns a specific record */
  event?: Maybe<EventRecord>;
  /** Returns a specific record */
  gallery?: Maybe<GalleryRecord>;
  /** Returns a specific record */
  join?: Maybe<JoinRecord>;
  /** Returns a specific record */
  partner?: Maybe<PartnerRecord>;
  /** Returns a specific record */
  podcast?: Maybe<PodcastRecord>;
  /** Returns a specific asset */
  upload?: Maybe<FileField>;
};
// ...
export type ResponsiveImage = {
  __typename?: 'ResponsiveImage';
  alt?: Maybe<Scalars['String']['output']>;
  aspectRatio: Scalars['FloatType']['output'];
  base64?: Maybe<Scalars['String']['output']>;
  bgColor?: Maybe<Scalars['String']['output']>;
  height: Scalars['IntType']['output'];
  sizes: Scalars['String']['output'];
  src: Scalars['String']['output'];
  srcSet: Scalars['String']['output'];
  title?: Maybe<Scalars['String']['output']>;
  webpSrcSet: Scalars['String']['output'];
  width: Scalars['IntType']['output'];
};
// ...
export enum SiteLocale {
  En = 'en',
  Pl = 'pl'
}
// ...
export type AboutQueryVariables = Exact<{
  locale?: InputMaybe<SiteLocale>;
}>;


export type AboutQuery = { __typename?: 'Query', about?: { __typename?: 'AboutRecord', description?: { __typename?: 'AboutModelDescriptionField', value: any } | null, malgosiaDescription?: { __typename?: 'AboutModelMalgosiaDescriptionField', value: any } | null, malgosiaImage?: { __typename?: 'FileField', responsiveImage?: { __typename?: 'ResponsiveImage', alt?: string | null, base64?: string | null, bgColor?: string | null, title?: string | null, aspectRatio: any, height: any, sizes: string, src: string, srcSet: string, webpSrcSet: string, width: any } | null } | null } | null };

export type EditionsQueryVariables = Exact<{
  locale?: InputMaybe<SiteLocale>;
  edition?: InputMaybe<Scalars['String']['input']>;
}>;


export type EditionsQuery = { __typename?: 'Query', event?: { __typename?: 'EventRecord', id: any, title?: string | null, edition?: string | null, new: any, date?: any | null, sign?: string | null, location?: string | null, lead: Array<{ __typename?: 'PersonRecord', name?: string | null, role?: string | null, image?: { __typename?: 'FileField', responsiveImage?: { __typename?: 'ResponsiveImage', alt?: string | null, base64?: string | null, bgColor?: string | null, title?: string | null, aspectRatio: any, height: any, sizes: string, src: string, srcSet: string, webpSrcSet: string, width: any } | null } | null }>, speakers: Array<{ __typename?: 'PersonRecord', name?: string | null, role?: string | null, image?: { __typename?: 'FileField', responsiveImage?: { __typename?: 'ResponsiveImage', alt?: string | null, base64?: string | null, bgColor?: string | null, title?: string | null, aspectRatio: any, height: any, sizes: string, src: string, srcSet: string, webpSrcSet: string, width: any } | null } | null }> } | null };

export type EventRoutesQueryVariables = Exact<{ [key: string]: never; }>;


export type EventRoutesQuery = { __typename?: 'Query', allEvents: Array<{ __typename?: 'EventRecord', edition?: string | null }> };

export type JoinUsQueryVariables = Exact<{
  locale?: InputMaybe<SiteLocale>;
}>;


export type JoinUsQuery = { __typename?: 'Query', allJoins: Array<{ __typename?: 'JoinRecord', social: Array<{ __typename?: 'SocialRecord', id: any, name?: string | null, description?: { __typename?: 'SocialModelDescriptionField', value: any } | null, icon?: { __typename?: 'FileField', responsiveImage?: { __typename?: 'ResponsiveImage', alt?: string | null, base64?: string | null, bgColor?: string | null, title?: string | null, aspectRatio: any, height: any, sizes: string, src: string, srcSet: string, webpSrcSet: string, width: any } | null } | null }> }> };

export class TypedDocumentString<TResult, TVariables>
  extends String
  implements DocumentTypeDecoration<TResult, TVariables>
{
  __apiType?: DocumentTypeDecoration<TResult, TVariables>['__apiType'];

  constructor(private value: string, public __meta__?: Record<string, any>) {
    super(value);
  }

  toString(): string & DocumentTypeDecoration<TResult, TVariables> {
    return this.value;
  }
}

export const AboutDocument = new TypedDocumentString(`
    query About($locale: SiteLocale) {
  about(locale: $locale) {
    description {
      value
    }
    malgosiaDescription {
      value
    }
    malgosiaImage {
      responsiveImage(imgixParams: {fit: crop, w: 300, h: 300, auto: format, q: 75}) {
        alt
        base64
        bgColor
        title
        aspectRatio
        height
        sizes
        src
        srcSet
        webpSrcSet
        width
      }
    }
  }
}
    `) as unknown as TypedDocumentString<AboutQuery, AboutQueryVariables>;
export const EditionsDocument = new TypedDocumentString(`
    query Editions($locale: SiteLocale, $edition: String) {
  event(locale: $locale, filter: {edition: {eq: $edition}}) {
    id
    title
    edition
    new
    date
    sign
    location
    lead {
      name
      role
      image {
        responsiveImage(imgixParams: {fit: crop, w: 300, h: 300, auto: format, q: 75}) {
          alt
          base64
          bgColor
          title
          aspectRatio
          height
          sizes
          src
          srcSet
          webpSrcSet
          width
        }
      }
    }
    speakers {
      name
      role
      image {
        responsiveImage(imgixParams: {fit: crop, w: 300, h: 300, auto: format, q: 75}) {
          alt
          base64
          bgColor
          title
          aspectRatio
          height
          sizes
          src
          srcSet
          webpSrcSet
          width
        }
      }
    }
  }
}
    `) as unknown as TypedDocumentString<EditionsQuery, EditionsQueryVariables>;
export const EventRoutesDocument = new TypedDocumentString(`
    query EventRoutes {
  allEvents {
    edition
  }
}
    `) as unknown as TypedDocumentString<EventRoutesQuery, EventRoutesQueryVariables>;
export const JoinUsDocument = new TypedDocumentString(`
    query JoinUs($locale: SiteLocale) {
  allJoins(locale: $locale) {
    social {
      id
      name
      description {
        value
      }
      icon {
        responsiveImage(imgixParams: {fit: crop, w: 300, h: 300, auto: format, q: 75}) {
          alt
          base64
          bgColor
          title
          aspectRatio
          height
          sizes
          src
          srcSet
          webpSrcSet
          width
        }
      }
    }
  }
}
    `) as unknown as TypedDocumentString<JoinUsQuery, JoinUsQueryVariables>;